:bug: refactor: remoção de classes e arquivos desnecessários

+++ api/src/main/java/com/app/service/CarroService.java
@@ -47,6 +47,10 @@ public class CarroService {
     }
 
     public void delete(int id) {
+        if (repository.findById(id) == null) {
+            throw new RuntimeException("Carro não encontrado");
+        }
+
         repository.delete(id);
     }
 }
